# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "version": "1.0.0"
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "consumes": [
        "multipart/form-data"
    ],
    "paths": {
        "/api/login": {
            "post": {
                "summary": "Connexion",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "password",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {}
                            },
                            "example": {}
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Deconnexion",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "password",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {}
                            },
                            "example": {}
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "summary": "Refresh",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "password",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {}
                            },
                            "example": {}
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Inscription",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "password",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "nom",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "prenom",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {}
                            },
                            "example": {}
                        }
                    }
                }
            }
        },
...
    }
}